Rename theme.ts
