+++ services/core-web/src/components/mine/Variances/MineVarianceTable.js
@@ -18,7 +18,7 @@ import { RED_CLOCK, EDIT_OUTLINE_VIOLET, TRASHCAN } from "@/constants/assets";
 import DocumentLink from "@/components/common/DocumentLink";
 import * as router from "@/constants/routes";
 import CoreTable from "@mds/common/components/common/CoreTable";
-import { getVarianceApplicationBadgeStatusType } from "@mds/common/constants/theme";
+import { getVarianceApplicationBadgeStatusType } from "@mds/common/constants/badgeStatusTypes";
 
 const propTypes = {
   handleVarianceSearch: PropTypes.func,
